+++ ui/main/shared/js/common.js
@@ -1003,11 +1003,6 @@ $(document).ready(function () {
     // now that loc has been updated, it's okay to show the page
     $('html').fadeIn(0).show();
 });
-(scene_mod_list['live_game'] = scene_mod_list['live_game'] || []).
-  unshift('coui://ui/mods/commander_regeneration/live_game.js');
-(scene_mod_list['live_game_unit_alert'] = scene_mod_list['live_game_unit_alert'] || []).
-  unshift('coui://ui/mods/commander_regeneration/live_game_unit_alert.js');
-
 (function() {
   var originalCall = engine.call
   engine.call = function(method) {